fix top score student and add docs
